Split selectors into function specific files
