+++ src/selectors/Bluetooth/sensorBlink.js
@@ -1,0 +1,13 @@
+/**
+ * mSupply Mobile
+ * Sustainable Solutions (NZ) Ltd. 2021
+ */
+
+import { selectBluetoothState } from './index';
+
+export const selectSendingBlinkTo = state => {
+  const bluetooth = selectBluetoothState(state);
+  const { blink } = bluetooth || {};
+  const { sendingBlinkTo = '' } = blink || {};
+  return sendingBlinkTo;
+};
