+++ src/selectors/Bluetooth/sensorDownload.js
@@ -1,30 +1,4 @@
-/**
- * mSupply Mobile
- * Sustainable Solutions (NZ) Ltd. 2021
- */
-
 import { selectBluetoothState } from './index';
-
-export const selectSendingBlinkTo = state => {
-  const bluetooth = selectBluetoothState(state);
-  const { blink } = bluetooth || {};
-  const { sendingBlinkTo = '' } = blink || {};
-  return sendingBlinkTo;
-};
-
-export const selectIsScanning = state => {
-  const bluetooth = selectBluetoothState(state);
-  const { scan } = bluetooth || {};
-  const { isScanning = false } = scan || {};
-  return isScanning;
-};
-
-export const selectScannedSensors = state => {
-  const bluetooth = selectBluetoothState(state);
-  const { scan } = bluetooth || {};
-  const { scannedSensorAddresses = [] } = scan || {};
-  return scannedSensorAddresses;
-};
 
 export const selectIsSyncingTemps = state => {
   const bluetooth = selectBluetoothState(state);
