+++ src/widgets/LastSensorDownload.js
@@ -14,7 +14,7 @@ import {
   selectIsDownloading,
   selectLastDownloadFailed,
   selectLastDownloadTime,
-} from '../selectors/Bluetooth/bluetooth';
+} from '../selectors/Bluetooth/sensorDownload';
 import { generalStrings } from '../localization';
 
 const formatLastSyncDate = date => (date ? moment(date).fromNow() : generalStrings.not_available);
